requirements.txt
Flask
pandas
requests
statsmodels
influxdb-client
plotly
pytest
python-dotenv